# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Bora Jogar API
  version: 0.1.0
paths:
...
  /api/v1/me/blocked-users:
    get: { summary: List private blocks, security: [{ cookieAuth: [] }], responses: { '200': { description: Blocked users } } }
  /api/v1/reports:
    post: { summary: File safety report, security: [{ cookieAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReportInput' } } } }, responses: { '201': { description: Report filed }, '422': { $ref: '#/components/responses/ValidationError' } } }
...
  /api/v1/me/delete:
    post: { summary: Request account deletion, security: [{ cookieAuth: [] }], responses: { '204': { description: Account deleted and sessions revoked } } }
...
  /api/v1/admin/reports:
    get: { summary: Review reports, security: [{ cookieAuth: [] }], responses: { '200': { description: Reports }, '403': { $ref: '#/components/responses/Forbidden' } } }
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: borajogar_session }
  schemas:
...
    ReportInput:
      type: object
      required: [category, description]
      properties:
        reportedUserId: { type: string, format: uuid, nullable: true }
        gameId: { type: string, format: uuid, nullable: true }
        category: { type: string, enum: [harassment, unsafe_behavior, repeated_no_show, false_profile, inappropriate_content, other] }
        description: { type: string, minLength: 1, maxLength: 2000 }
        blockReportedUser: { type: boolean }
...
    ApiError:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
        requestId: { type: string }
  responses:
...
    Forbidden: { description: Administrator access required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }